requirements.txt
# Core ML and Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Framework and API
streamlit>=1.28.0
streamlit-folium>=0.15.0
folium>=0.15.0
fastapi>=0.104.0
uvicorn>=0.24.0
slowapi>=0.1.9

# Testing and Quality
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# ML Operations and Monitoring
mlflow>=2.8.0
evidently>=0.4.0
shap>=0.43.0
optuna>=3.4.0
category-encoders>=2.6.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
joblib>=1.3.0
tqdm>=4.66.0

# Development and Deployment
docker>=6.1.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: For advanced monitoring
prometheus-client>=0.19.0
grafana-api>=1.0.3
